Create Building a document analyser
This code builds a lambda function that is triggered by an event, then uses textract to extract text from a document and amazon medical comprehend to perform a medical entity analysis. It also handles errors using process error and writes the analyst's results into a CSV file that is stored in a separate S3 bucket.